requirements.txt
prophet
numpy
streamlit==1.8.1
pandas==1.4.1
yfinance==0.1.70
plotly
